+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
-## v0.1.1
+## v0.2.0 (2020-05-08)
 
-+ Modify the export method
++ Add `DOMElement.insertBefore`.
+
+## v0.1.1 (2020-05-07)
+
++ Modify the export method.